requirements.txt
aiosqlite==0.22.1
fastapi==0.136.0
httpx==0.28.1
pydantic==2.13.3
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
sqlalchemy==2.0.49
uvicorn==0.45.0
